+++ R/get_power.R
@@ -1,150 +1,96 @@
 
-#'Get NASA-POWER Data and Return a Tidy Data Frame
+#'@title Get NASA-POWER Data and Return a Tidy Data Frame
 #'
-#'@description Download NASA-POWER (Prediction of Worldwide Energy Resource)
-#'  data for a given time span.
+#'@description Get NASA-POWER data and return a tidy data frame. All options
+#'offered by the official NASA-POWER API are supported.
 #'
-#'@param name Data set name, currently supported are \code{'AG'}, \code{'SB'}
-#'  and \code{'SSE'}. Defaults to \code{'AG'}.  See Details.
-#'@param latlon Geographic coordinates for a cell or region as x, y coordinates
-#'  or \code{"Global"}.  Defaults to \code{"Global"}.
-#'@param vars Solar or meteorological variables to download. See Details for
-#'  \code{vars}
-#'@param dates A length-2 character vector of start and end dates in that order,
-#'  \emph{e.g.}, \code{dates = c("1983-01-01", "2017-12-31")}.  If \code{date}
-#'  is unspecified,  defaults to a start date of "1983-01-01" (the earliest
-#'  available data) and an end date of current date.
-#'@param time_step Temporal averages, currently supported are \code{'DAILY'},
-#'  \code{'INTERANNUAL'} and \code{'CLIMATOLOGY'}.  Defaults to \code{'DAILY'}.
-#'  See Details.
+#'@export
+#'@param community Dataset name, currently supported are AG, SB and SSE. See
+#'  argument details for more.
+#'@param lonlat A numeric vector of geographic coordinates for a cell or region
+#'  entered as x, y coordinates in that order, *e.g.*, `lonlat = c(36, 45)`.
+#'  Leave unspecified for global coverage.  See argument details for more.
+#'@param pars A character vector of solar or meteorological variables to
+#'  download.  See the `value` field of [`parameters`] for a full list of valid
+#'  values or visit the [POWER website](https://power.larc.nasa.gov/new/) for
+#'  the Parameter Dictionary.
+#'@param dates A character vector of start and end dates in that order, *e.g.*,
+#'  `dates = c("1983-01-01", "2017-12-31")`.  See argument details for more.
+#'@param temporal_average Temporal average for data being queried, currently
+#'  supported are DAILY, INTERANNUAL, CLIMATOLOGY.  See argument details for
+#'  more.
 #'
-#'@return A tidy \code{\link[base]{data.frame}} object of the requested
-#'  variable(s) for the requested longitude and latitude values.
+#'@details Further details for each of the arguments are provided in their
+#'respective sections following below.
 #'
-#'@section Argument details for \code{name}:
-#'\describe{
-#'  \item{\code{AG}}{Provides access to the Agroclimatology Archive, which
+#'@section Argument details for `community`: There are three valid values, one
+#'  must be supplied. This  will affect the units of the parameter and the
+#'  temporal display of time series data.
+#'
+#'  \describe{ \item{AG}{Provides access to the Agroclimatology Archive, which
 #'  contains industry-friendly parameters formatted for input to crop models.}
-#'  \item{\code{SB}}{Provides access to the Sustainable Buildings Archive, which
+#'
+#'  \item{SB}{Provides access to the Sustainable Buildings Archive, which
 #'  contains industry-friendly parameters for the buildings community to include
 #'  parameters in multi-year monthly averages.}
-#'  \item{\code{SSE}}{Provides access to the Renewable Energy Archive, which
-#'  contains parameters specifically tailored to assist in the design of solar
-#'  and wind powered renewable energy systems.}
-#'  }
 #'
-#'@section Argument details for \code{time_step}:
-#' \describe{
-#'  \item{DAILY}{Daily average by year.}
-#'  \item{INTERANNUAL}{Monthly and annual average by year.}
-#'  \item{CLIMATOLOGY}{Long term monthly averages.} }
+#'  \item{SSE}{Provides access to the Renewable Energy Archive, which contains
+#'  parameters specifically tailored to assist in the design of solar and wind
+#'  powered renewable energy systems.} }
 #'
-#'@section Argument details for \code{latlon}:
-#' To get a specific cell, 1/2 x 1/2 degree, supply a length-2 numeric vector
-#' giving the decimal degree longitude and latitude in that order for data to
-#' download, \emph{e.g.}, \code{lonlat = c(-179.5, -89.5)}.
+#'@section Argument details for `lonlat`: \describe{ \item{For a single
+#'  point}{To get a specific cell, 1/2 x 1/2 degree, supply a length-2 numeric
+#'  vector giving the decimal degree longitude and latitude in that order for
+#'  data to download, *e.g.*, `lonlat = c(-179.5, -89.5)`.}
 #'
-#' To get a region, supply a length-4 numeric vector as \code{xmin, xmax, ymin,
-#' ymax} in that order for a given region, \emph{e.g.}, a bounding box for
-#' the southwestern corner of Australia:
-#' \code{lonlat = c(112.5, 122.5, -55.5, -45.5)}. \emph{Max bounding box is 10 x
-#' 10 degrees of 1/2 x 1/2 degree data}, \emph{i.e.}, 100 Points Max Total.
+#'  \item{For regional coverage}{To get a region, supply a length-4 numeric
+#'  vector as `lonlat = c(xmin, xmax, ymin, ymax)` in that order for a given
+#'  region, *e.g.*, a bounding box for the southwestern corner of Australia:
+#'  `lonlat = c(112.5, 122.5, -55.5, -45.5)`. *Max bounding box is 10 x 10
+#'  degrees* of 1/2 x 1/2 degree data, *i.e.*, 100 points maximum in total.}
 #'
-#' To download a complete surface for the entire globe, set
-#' \code{lonlat = "Global"}.
+#'  \item{For global coverage}{Not used.} }
 #'
-#'@note The order in which the \code{vars} are listed will be the order of the
-#'  columns in the data frame that \code{get_power()} returns.
+#'@section Argument details for `dates`: If `dates` is unspecified, defaults to
+#'  a start date of 1983-01-01 (the earliest available data) and an end date of
+#'  current date according to the system.
 #'
-#' @examples
-#' \dontrun{
-#' nasa <- get_power(lonlat = c(-179.5, -89.5))
-#' }
+#'  If one date only is provided, it will be treated as both the start date and
+#'  the end date and only a single day's values will be returned.
+#'
+#'@section Argument details for `temporal_average`: There are three valid
+#'  values, one must be supplied. \describe{ \item{DAILY}{The daily average of
+#'  `pars` by year.} \item{INTERANNUAL}{The monthly average of `pars` by year.}
+#'  \item{CLIMATOLOGY}{The monthly average of `pars` at the surface of the earth
+#'  for a given month, averaged for that month over the 30-year period (Jan.
+#'  1984 - Dec. 2013).} }
 #'
 #'@references
 #'\url{https://power.larc.nasa.gov/documents/Agroclimatology_Methodology.pdf}
 #'@author Adam H. Sparks, adamhsparks@gmail.com
 #'
-#'@export
-get_power <-
-  function(name = "AG",
-           lonlat = "Global",
-           vars = NULL,
-           dates = c("1983-01-01", as.character(Sys.Date())),
-           time_step = "DAILY") {
-    #check user inputs, see internal_functions.R for these
+get_power <- function(community = NULL,
+                      lonlat = NULL,
+                      pars = NULL,
+                      dates = NULL,
+                      temporal_average = NULL) {
 
-    identifier <- .check_lonlat_cell(lonlat)
-    .check_vars(vars, name)
-    dates <- .check_dates(dates)
+  # user input checks and formatting -------------------------------------------
+  # see internal_functions.R for these functions
+  dates <- check_dates(dates)
+  lonlat_identifier <- check_lonlat(lonlat, pars)
+  pars <- check_pars(pars)
+  community <- check_community(community)
+  temporal_average <- check_tempavg(temporal_average)
 
-    # check if website is responding
-    power_url <-
-      "asdc-arcgis.larc.nasa.gov/cgi-bin/power/v1beta/DataAccess.py?"
-    .check_response(power_url)
+  # submit query ---------------------------------------------------------------
+  # see internal_functions.R for this function
+  NASA <- get_power(community,
+                    lonlat_identifier,
+                    pars,
+                    dates,
+                    temporal_average)
 
-    if (isTRUE(any(stringi::stri_detect_fixed(vars, "RAIN")))) {
-      if (start_date < "1997-01-01") {
-        message("POWER does not supply precipitation data before 1997-01-01.\n")
-      }
-    }
 
-    # create the query list ----------------------------------------------------
-
-    # concatenate all the download vars into a single string for use below
-    download_vars <- paste0("&p=", vars, collapse = "")
-
-    # assemble the query list, previous step is necessary because httr::query
-    # seems to remove duplicated list item names, all vars = "p" so...
-    power_query <- list(
-
-      identifier = identifier,
-      parameters = paste0(vars, collapse = ","),
-      startDate = dates[[1]],
-      endDate = dates[[2]],
-      lon = lonlat[1],
-      lat = lonlat[2],
-      request = "execute"
-    )
-
-    # submit query -------------------------------------------------------------
-    # see internal-functions for this function
-    NASA <- .get_NASA(power_url, download_vars, power_query)
-
-    # create data frame of downloaded data -------------------------------------
-    # colnames first
-    colnames <-
-      unlist(strsplit(NASA[grep("-END HEADER-", NASA) - 1], split = " "))
-
-    # clean up column names, otherwise there are empty values in the vector
-    colnames <- unique(colnames[colnames != ""])
-
-    NASA <- utils::read.table(
-      text = NASA,
-      skip = grep("-END HEADER-", NASA),
-      na.strings = "-",
-      nrows = as.numeric(dates[[2]] - dates[[1]]) + 1,
-      stringsAsFactors = FALSE
-    )
-
-    # check df, if all NA, stop, if has data, return df
-    .check_nasa_df(NASA)
-
-    names(NASA) <- colnames
-
-    # create a tidy data frame object
-    NASA["LON"] <- lonlat[1]
-    NASA["LAT"] <- lonlat[2]
-
-    # add additional date fields
-    NASA["YYYYMMDD"] <- as.Date(NASA$DOY, origin = dates[[1]] - 1)
-    NASA["MONTH"] <- format(as.Date(NASA$YYYYMMDD), "%m")
-    NASA["DAY"] <- format(as.Date(NASA$YYYYMMDD), "%d")
-
-    # rearrange columns
-    refcols <-
-      c("YEAR", "MONTH", "DAY", "YYYYMMDD", "DOY", "LON", "LAT")
-    NASA <- NASA[, c(refcols, setdiff(names(NASA), refcols))]
-
-    return(NASA)
-  }
+  return(NASA)
+}